+++ python/mxnet/kvstore/kvstore.py
@@ -29,50 +29,9 @@ from ..base import check_call, string_types, mx_uint, py_str
 from ..base import NDArrayHandle, KVStoreHandle
 from .. import optimizer as opt
 from ..profiler import set_kvstore_handle
+from .base import _ctype_key_value, _ctype_dict, AbstractKVStore
 
-__all__ = ['KVStore', 'create']
-
-def _ctype_key_value(keys, vals):
-    """Returns ctype arrays for the key-value args, and the whether string keys are used.
-    For internal use only.
-    """
-    if isinstance(keys, (tuple, list)):
-        assert(len(keys) == len(vals))
-        c_keys = []
-        c_vals = []
-        use_str_keys = None
-        for key, val in zip(keys, vals):
-            c_key_i, c_val_i, str_keys_i = _ctype_key_value(key, val)
-            c_keys += c_key_i
-            c_vals += c_val_i
-            use_str_keys = str_keys_i if use_str_keys is None else use_str_keys
-            assert(use_str_keys == str_keys_i), "inconsistent types of keys detected."
-        c_keys_arr = c_array(ctypes.c_char_p, c_keys) if use_str_keys \
-                     else c_array(ctypes.c_int, c_keys)
-        c_vals_arr = c_array(ctypes.c_void_p, c_vals)
-        return (c_keys_arr, c_vals_arr, use_str_keys)
-
-    assert(isinstance(keys, (int,) + string_types)), \
-           "unexpected type for keys: " + str(type(keys))
-    use_str_keys = isinstance(keys, string_types)
-    if isinstance(vals, NDArray):
-        c_keys = c_str_array([keys]) if use_str_keys \
-                 else c_array_buf(ctypes.c_int, array('i', [keys]))
-        return (c_keys, c_handle_array([vals]), use_str_keys)
-    else:
-        for value in vals:
-            assert(isinstance(value, NDArray))
-        c_keys = c_str_array([keys] * len(vals)) if use_str_keys \
-                 else c_array_buf(ctypes.c_int, array('i', [keys] * len(vals)))
-        return (c_keys, c_handle_array(vals), use_str_keys)
-
-def _ctype_dict(param_dict):
-    """Returns ctype arrays for keys and values(converted to strings) in a dictionary"""
-    assert(isinstance(param_dict, dict)), \
-        "unexpected type for param_dict: " + str(type(param_dict))
-    c_keys = c_array(ctypes.c_char_p, [c_str(k) for k in param_dict.keys()])
-    c_vals = c_array(ctypes.c_char_p, [c_str(str(v)) for v in param_dict.values()])
-    return (c_keys, c_vals)
+__all__ = ['KVStore']
 
 def _updater_wrapper(updater):
     """A wrapper for the user-defined handle."""
@@ -93,124 +52,6 @@ def _get_kvstore_server_command_type(command):
     assert (command in command_types), "Unknown command type to send to server"
     return command_types[command]
 
-class AbstractKVStore(object):
-    """An abstract key-value store interface for data parallel training."""
-
-    def broadcast(self, key, value, out=None):
-        """ Broadcast the value NDArray at rank 0 to all ranks' out. If out is None,
-        the result is stored in `value`.
-
-        Parameters
-        ----------
-        key : str, int, or sequence of str or int
-            Keys.
-
-        value : NDArray, list of NDArray, or list of list of NDArray
-            Values corresponding to the keys.
-
-        out : optional NDArray, list of NDArray, or list of list of NDArray
-            Values corresponding to the keys.
-        """
-        raise NotImplementedError()
-
-    def pushpull(self, key, value, out=None, priority=0):
-        """ Performs push and pull a single value or a sequence of values from the store.
-
-        This function is coalesced form of push and pull operations.
-
-        `value` is pushed to the kvstore server for the specified keys and the updated
-        values are pulled from the server to `out`. If `out` is not specified the pulled
-        values are written to `value`.
-
-        Parameters
-        ----------
-        key : str, int, or sequence of str or int
-            Keys.
-
-        value : NDArray, list of NDArray, or list of list of NDArray
-            Values corresponding to the keys.
-
-        out: NDArray or list of NDArray or list of list of NDArray
-            Values corresponding to the keys.
-
-        priority : int, optional
-            The priority of the pull operation.
-            Higher priority pull operations are likely to be executed before
-            other pull actions.
-        """
-        raise NotImplementedError()
-
-    def set_optimizer(self, optimizer):
-        """ Registers an optimizer with the kvstore.
-
-        When using a single machine, this function updates the local optimizer.
-        If using multiple machines and this operation is invoked from a worker node,
-        it will serialized the optimizer with pickle and send it to all servers.
-        The function returns after all servers have been updated.
-
-        Parameters
-        ----------
-        optimizer : Optimizer
-            The new optimizer for the store
-        """
-        raise NotImplementedError()
-
-    def save_optimizer_states(self, fname, dump_optimizer=False):
-        """Saves the optimizer (updater) state to a file. This is often used when checkpointing
-        the model during training.
-
-        Parameters
-        ----------
-        fname : str
-            Path to the output states file.
-        dump_optimizer : bool, default False
-            Whether to also save the optimizer itself. This would also save optimizer
-            information such as learning rate and weight decay schedules.
-        """
-        raise NotImplementedError()
-
-    def load_optimizer_states(self, fname):
-        """Loads the optimizer (updater) state from the file.
-
-        Parameters
-        ----------
-        fname : str
-            Path to input states file.
-        """
-        raise NotImplementedError()
-
-    @property
-    def type(self):
-        """ Returns the type of this kvstore.
-
-        Returns
-        -------
-        type : str
-            the string type
-        """
-        raise NotImplementedError()
-
-    @property
-    def rank(self):
-        """ Returns the rank of this worker node.
-
-        Returns
-        -------
-        rank : int
-            The rank of this node, which is in range [0, num_workers())
-        """
-        raise NotImplementedError()
-
-    @property
-    def num_workers(self):
-        """Returns the number of worker nodes.
-
-        Returns
-        -------
-        size :int
-            The number of worker nodes.
-        """
-        raise NotImplementedError()
 
 class KVStore(AbstractKVStore):
     """A key-value store for synchronization of values, over multiple devices."""
@@ -849,47 +690,3 @@ class KVStore(AbstractKVStore):
         """
         check_call(_LIB.MXKVStoreSendCommmandToServers(
             self.handle, mx_uint(head), c_str(body)))
-
-def create(name='local'):
-    """Creates a new KVStore.
-
-    For single machine training, there are two commonly used types:
-
-    ``local``: Copies all gradients to CPU memory and updates weights there.
-
-    ``device``: Aggregates gradients and updates weights on GPUs. With this setting,
-    the KVStore also attempts to use GPU peer-to-peer communication,
-    potentially accelerating the communication.
-
-    For distributed training, KVStore also supports a number of types:
-
-    ``dist_sync``: Behaves similarly to ``local`` but with one major difference.
-    With ``dist_sync``, batch-size now means the batch size used on each machine.
-    So if there are ``n`` machines and we use batch size ``b``,
-    then ``dist_sync`` behaves like ``local`` with batch size ``n * b``.
-
-    ``dist_device_sync``: Identical to ``dist_sync`` with the difference similar
-    to ``device`` vs ``local``.
-
-    ``dist_async``: Performs asynchronous updates.
-    The weights are updated whenever gradients are received from any machine.
-    No two updates happen on the same weight at the same time. However, the order is not
-    guaranteed.
-
-    Parameters
-    ----------
-    name : {'local', 'device', 'nccl', 'dist_sync', 'dist_device_sync', 'dist_async'}
-        The type of KVStore.
-    Returns
-    -------
-    kv : KVStore
-        The created KVStore.
-    """
-    if not isinstance(name, string_types):
-        raise TypeError('name must be a string')
-    handle = KVStoreHandle()
-    check_call(_LIB.MXKVStoreCreate(c_str(name),
-                                    ctypes.byref(handle)))
-    kv = KVStore(handle)
-    set_kvstore_handle(kv.handle)
-    return kv
